logrotate: add new struct type and rotation handling
The struct 'LogRotateOptions' stores option values set for the current
affile destination driver.
Logrotation is performed in the function 'do_logrotate' and for now
implements a simple size based rotation.

There is no in-memory storage or list to keep track of rotated files.
With each rotation the number of log files is established through the
filenames.