logwriter: signal logrotation to logpipe after final flush
The logwriter thread notifies its controlling logpipe instance
(AFFileDestWriter) that it is done writing to the logfile (for now)
and logrotation can be performed, if pending.

After the files have been rotated the current logfile needs to be
reopened. However a call to affile_dw_reopen from within the worker
threads log_writer_work_perform function would lead to a deadlock (the
thread waits for itself to finish).
Therefore the thread updates its LogProtoClient with the fd set by the
AFFileDestWriter through log_pipe_notify.

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>