affile: add filesize caching to log_writer_flush
Within every call to 'log_writer_flush' the current filesize is retrieve
via stat. While emptying the message queue the function
'log_writer_write_message' reports back the actual message size which is
used to keep track of the filesize.
After a message has been taken out of the message queue and being ready
to be flushed, a call via 'log_pipe_notify' informs the associated
AFFileDestWriter to check is logroation is necessary based on the cached
filesize.

Signed-off-by: Alexander Zikulnig <42011733+azike@users.noreply.github.com>